requirements.txt
Pillow>=9.3
PyMuPDF==1.21.1